# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the project",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "description": "Returns list of products",
                "operationId": "getProductsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string"
                                        },
                                        "from": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "last_page_url": {
                                            "type": "string"
                                        },
                                        "next_page_url": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "prev_page_url": {
                                            "type": "string"
                                        },
                                        "to": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product information",
                "description": "Returns product data",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "required": [
                    "name",
                    "price",
                    "category_id",
                    "brand_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the product",
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "description": "Category ID",
                        "type": "integer",
                        "example": 1
                    },
                    "brand_id": {
                        "description": "Brand ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Product name",
                        "type": "string",
                        "example": "Product name"
                    },
                    "slug": {
                        "description": "Product slug",
                        "type": "string",
                        "example": "product-name"
                    },
                    "images": {
                        "description": "Product images",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "image1.jpg",
                            "image2.jpg"
                        ]
                    },
                    "description": {
                        "description": "Product description",
                        "type": "string",
                        "example": "Product description"
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "is_active": {
                        "description": "Is product active",
                        "type": "boolean",
                        "example": true
                    },
                    "is_featured": {
                        "description": "Is product featured",
                        "type": "boolean",
                        "example": false
                    },
                    "in_stock": {
                        "description": "Is product in stock",
                        "type": "boolean",
                        "example": true
                    },
                    "on_sale": {
                        "description": "Is product on sale",
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "description": "Created at timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Updated at timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}